requirements.txt
flask
python-dotenv
bs4
requests